pyproject.toml
[tool.poetry]
name = "Flight App"
version = "0.1.0"
description = "Flight App API"
authors = ["marius.crisan <marius.crisan.tech@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
faker = "^22.0.0"
django-password-validators = "^1.7.3"
django-redis = "^5.4.0"
django-debug-toolbar = "^4.3.0"
django-health-check = "^3.18.1"
celery = {extras = ["redis"], version = "^5.3.6"}
pandas = "^2.2.1"
geopy = "^2.4.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
